+++ packages/button/src/__tests__/FakeButton.tsx
@@ -17,7 +17,6 @@ const themes: ButtonTheme[] = [
   "secondary",
   "warning",
   "error",
-  "default",
 ];
 const themeTypes: ButtonThemeType[] = ["flat", "contained", "outline"];
 const buttonTypes: ButtonType[] = ["text", "icon"];
@@ -95,6 +94,7 @@ describe("FakeButton", () => {
     expect(onClick).toBeCalled();
   });
 
+  // this is actually really bad practice and kind of worthess
   it("should render correctly based on the theme props", () => {
     flattenDeep(
       buttonTypes.map(buttonType =>